requirements.txt
anyio==3.5.0
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
fastapi==0.72.0
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.28
protobuf==3.19.3
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.0

jwt~=1.3.1
pandas~=1.4.0
PyJWT~=2.3.0